# apps/web/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DyPos CL API v1",
    "description": "API REST del sistema DyPos CL. Requiere autenticación NextAuth (cookie de sesión). Rate limit: 100 req/min por IP.",
    "version": "1.0.0",
    "contact": { "name": "DyPos CL" }
  },
  "servers": [
    { "url": "/api/v1", "description": "Producción / Desarrollo local" }
  ],
  "security": [{ "cookieAuth": [] }],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authjs.session-token"
      }
    },
    "schemas": {
...
      "DetalleVenta": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "cantidad": { "type": "integer", "example": 2 },
          "precioUnitario": { "type": "integer", "example": 990 },
          "subtotal": { "type": "integer", "example": 1980 },
          "productoId": { "type": "integer" },
          "producto": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "nombre": { "type": "string" }
            }
          }
        }
      },
      "Venta": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "numeroBoleta": { "type": "string", "example": "BOL-1729123456789" },
          "publicToken": { "type": "string", "description": "Token público no enumerable para compartir comprobante interno" },
          "fecha": { "type": "string", "format": "date-time" },
          "subtotal": { "type": "integer", "description": "Subtotal bruto en CLP (antes de descuentos)" },
          "descuentoPct": { "type": "number", "description": "Descuento porcentual aplicado (0-100)", "minimum": 0, "maximum": 100, "default": 0 },
          "descuentoMonto": { "type": "integer", "description": "Descuento fijo en CLP (sobre base tras descuento %)", "minimum": 0, "default": 0 },
          "impuesto": { "type": "integer", "description": "IVA 19% sobre base imponible (tras descuentos), CLP" },
          "total": { "type": "integer", "description": "base imponible + IVA, CLP" },
          "metodoPago": {
            "type": "string",
            "enum": ["EFECTIVO", "DEBITO", "CREDITO", "TRANSFERENCIA"]
          },
          "clienteId": { "type": "integer", "nullable": true },
          "usuarioId": { "type": "integer" },
          "cliente": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": { "type": "integer" },
              "nombre": { "type": "string" },
              "rut": { "type": "string" }
            }
          },
          "usuario": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "nombre": { "type": "string" }
            }
          },
          "detalles": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/DetalleVenta" }
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "No autenticado" }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "401 — No autenticado",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
      "BadRequest": {
        "description": "400 — Datos de entrada inválidos",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "RateLimit": {
        "description": "429 — Rate limit excedido (100 req/min)",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      }
    }
  },
  "paths": {
...
    "/ventas": {
      "get": {
        "summary": "Listar ventas con filtros",
        "tags": ["Ventas"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } },
          { "name": "desde", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "YYYY-MM-DD" },
          { "name": "hasta", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "YYYY-MM-DD" },
          { "name": "clienteId", "in": "query", "schema": { "type": "integer" } },
          { "name": "metodoPago", "in": "query", "schema": { "type": "string", "enum": ["EFECTIVO", "DEBITO", "CREDITO", "TRANSFERENCIA"] } }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de ventas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Venta" } },
                    "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimit" }
        }
      },
      "post": {
        "summary": "Crear venta (transacción atómica: actualiza stock + ventas de productos + compras del cliente)",
        "tags": ["Ventas"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["items"],
                "properties": {
                  "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                      "type": "object",
                      "required": ["productoId", "cantidad"],
                      "properties": {
                        "productoId": { "type": "integer", "minimum": 1 },
                        "cantidad": { "type": "integer", "minimum": 1 }
                      }
                    }
                  },
                  "clienteId": { "type": "integer", "description": "Opcional" },
                  "metodoPago": {
                    "type": "string",
                    "enum": ["EFECTIVO", "DEBITO", "CREDITO", "TRANSFERENCIA"],
                    "default": "EFECTIVO"
                  },
                  "descuentoPct": {
                    "type": "number",
                    "description": "Descuento porcentual (0-100). Se aplica antes del descuento fijo. Valores fuera de rango son clamped.",
                    "minimum": 0,
                    "maximum": 100,
                    "default": 0
                  },
                  "descuentoMonto": {
                    "type": "integer",
                    "description": "Descuento fijo en CLP sobre la base tras %. Si excede la base, se clampa (no genera base negativa).",
                    "minimum": 0,
                    "default": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Venta creada" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "description": "Stock insuficiente o producto inactivo" },
          "429": { "$ref": "#/components/responses/RateLimit" }
        }
      }
    },
...
  }
}